fix: Board Profile 관련 누락 매서드 추가

+++ a406_backend/src/main/java/com/social/a406/domain/board/dto/BoardProfileResponse.java
@@ -9,7 +9,7 @@ public class BoardProfileResponse {
     private Long BoardId;
     private Double x;
     private Double y;
-    private Integer z;
+    private Double z;
     private String keyword;
     private Integer pageNumber;
     private String imageUrl;
